+++ libs/qt/source/ftrack_qt/widgets/accordion/accordion_widget.py
@@ -17,14 +17,22 @@ class AccordionBaseWidget(QtWidgets.QFrame):
     doubleClicked = QtCore.Signal(
         object
     )  # Emitted when accordion is double clicked
+    bin_clicked = QtCore.Signal(object)  # Emitted when bin icon is clicked
     show_options_overlay = QtCore.Signal(object)
     hide_options_overlay = QtCore.Signal()
     enabled_changed = QtCore.Signal(object)
+    title_changed = QtCore.Signal(object)
+    title_edited = QtCore.Signal(object)
 
     @property
     def title(self):
         '''Return the title text shown in header by default'''
         return self._title
+
+    @property
+    def editable_title(self):
+        '''Return the title text shown in header by default'''
+        return self._editable_title
 
     @property
     def selectable(self):
@@ -68,16 +76,29 @@ class AccordionBaseWidget(QtWidgets.QFrame):
         '''(Checkable) Return True if checked'''
         return self._checked
 
+    @property
+    def removable(self):
+        '''
+        Return True if accordion is removable - can be deleted by user
+        '''
+        return self._removable
+
+    @property
+    def previous_title(self):
+        return self._header_widget.previous_title
+
     def __init__(
         self,
         selectable=False,
         show_checkbox=False,
         checkable=False,
         title=None,
+        editable_title=False,
         selected=False,
         checked=True,
         collapsable=True,
         collapsed=True,
+        removable=False,
         parent=None,
     ):
         '''
@@ -103,7 +124,9 @@ class AccordionBaseWidget(QtWidgets.QFrame):
         self._checkable = checkable
         self._show_checkbox = show_checkbox
         self._title = title
+        self._editable_title = editable_title
         self._collapsable = collapsable
+        self._removable = removable
 
         self._selected = selected
         self._checked = checked
@@ -140,11 +163,13 @@ class AccordionBaseWidget(QtWidgets.QFrame):
         # Create Header
         self._header_widget = AccordionHeaderWidget(
             title=self.title,
+            editable_title=self.editable_title,
             checkable=self.checkable,
             checked=self.checked,
             show_checkbox=self.show_checkbox,
             collapsable=self.collapsable,
             collapsed=self.collapsed,
+            removable=self.removable,
         )
 
         # Add header to main widget
@@ -170,12 +195,15 @@ class AccordionBaseWidget(QtWidgets.QFrame):
         self._header_widget.arrow_clicked.connect(
             self._on_header_arrow_clicked
         )
+        self._header_widget.bin_clicked.connect(self._on_header_bin_clicked)
         self._header_widget.show_options_overlay.connect(
             self._on_show_options_overlay_callback
         )
         self._header_widget.hide_options_overlay.connect(
             self._on_hide_options_overlay_callback
         )
+        self._header_widget.title_changed.connect(self._on_title_changed)
+        self._header_widget.title_edited.connect(self._on_title_edited)
         self._content_widget.setVisible(not self._collapsed)
         self._content_widget.setEnabled(self.checked)
 
@@ -184,6 +212,12 @@ class AccordionBaseWidget(QtWidgets.QFrame):
 
     def _on_hide_options_overlay_callback(self):
         self.hide_options_overlay.emit()
+
+    def _on_title_changed(self, title):
+        self.title_changed.emit(title)
+
+    def _on_title_edited(self, title):
+        self.title_edited.emit(title)
 
     def add_option_widget(self, widget, section_name):
         self._header_widget.add_option_widget(widget, section_name)
@@ -234,6 +268,10 @@ class AccordionBaseWidget(QtWidgets.QFrame):
             # This is the way to collapse
             self.toggle_collapsed()
 
+    def _on_header_bin_clicked(self, event):
+        '''Callback on header arrow user click'''
+        self.bin_clicked.emit(event)
+
     def toggle_collapsed(self):
         '''Toggle the accordion collapsed state'''
         self._collapsed = not self._collapsed
@@ -273,6 +311,10 @@ class AccordionBaseWidget(QtWidgets.QFrame):
                 'background',
                 'selected' if self._selected else 'transparent',
             )
+
+    def set_title(self, new_title):
+        self._title = new_title
+        self._header_widget.set_title(self.title)
 
     def teardown(self):
         '''Teardown the header widget - properly cleanup the options overlay'''