+++ libs/constants/release_notes.md
@@ -1,5 +1,10 @@
 # ftrack Constants library release Notes
 
+## Upcoming
+
+* [new] Add HOST_SYNC_TOOL_CONFIG_TOPIC constant.
+
+
 ## v2.0.0
 2024-02-12